fix: resolve 6 compilation errors in v2 feature code
- AssistantCustomizeViewModel: updateSettings() -> saveSettings()
- LiveCallActivity: fix OrchestratorState import path (remove .Companion)
- VoicemailAdapter: fix formatDuration parameter type Int vs Long
- VoicemailViewModel: pass voicemail.id instead of object to deleteVoicemail()
- CallOrchestratorService: fix missing string resource assistant_active_title
- VoicemailManager: fix missing string resource voicemail_new_notification_title
- Add gradle wrapper scripts (gradlew, gradlew.bat, wrapper jar)

+++ app/src/main/java/com/suanmesgulum/app/presentation/voicemail/VoicemailAdapter.kt
@@ -46,7 +46,7 @@ class VoicemailAdapter(
         holder.itemView.setOnClickListener { onItemClick(item) }
     }
 
-    private fun formatDuration(seconds: Long): String {
+    private fun formatDuration(seconds: Int): String {
         val min = seconds / 60
         val sec = seconds % 60
         return String.format("%d:%02d", min, sec)
